requirements.txt
anthropic>=0.40.0
deepagents
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-anthropic
langsmith>=0.1.99
langgraph>=1.0,<2.0
kubernetes>=29.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
